+++ skelet3d/libfixer.py
@@ -2,9 +2,7 @@
 # -*- coding: utf-8 -*-
 
 
-import logging
-
-logger = logging.getLogger(__name__)
+from loguru import logger
 
 import zipfile
 import glob